requirements.txt
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
colorama==0.4.6
cryptography==41.0.4
dataclasses==0.6
defusedxml==0.8.0rc2
deprecation==2.1.0
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.2.0
django-filter==23.3
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.93.4
exponent-server-sdk==2.0.0
gotrue==1.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==2.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
pluggy==1.3.0
postgrest==0.13.0
postgrest-py==0.4.0
psycopg2-binary==2.9.9
py==1.11.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
realtime==1.0.0
realtime-py==0.1.3
requests==2.25.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rollbar==1.0.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
storage3==0.6.1
StrEnum==0.4.15
supabase==2.0.3
supabase-py==0.0.2
supafunc==0.3.1
toml==0.10.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
websockets==10.4
